requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
pymysql==1.1.0
gunicorn==21.2.0
